# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aletheia Deep Research API",
    "description": "\n    ## 🔍 API para análisis e investigación profunda\n    \n    Esta API proporciona capacidades avanzadas de investigación utilizando:\n    \n    - **Investigación Iterativa**: Búsqueda profunda con múltiples iteraciones\n    - **Procesamiento Paralelo**: Optimizado para alto rendimiento\n    - **Trazabilidad Completa**: Monitoreo y observabilidad con OpenTelemetry\n    - **Múltiples Fuentes**: Integración con Tavily y Saptiva APIs\n    \n    ### 📋 Casos de Uso\n    \n    - Análisis de mercado y competencia\n    - Investigación académica y científica  \n    - Due diligence empresarial\n    - Monitoreo de tendencias\n    \n    ### 🚀 Características\n    \n    - ✅ Auto-escalado en Kubernetes\n    - ✅ CI/CD enterprise-grade\n    - ✅ Seguridad y compliance\n    - ✅ Métricas de calidad\n    ",
    "contact": {
      "name": "Aletheia Development Team",
      "url": "https://github.com/saptiva-ai/alethia_deepresearch",
      "email": "dev@saptiva.ai"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/research": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Iniciar investigación estándar",
        "description": "Inicia una nueva tarea de investigación con procesamiento paralelo optimizado.\n    \n    - **Procesamiento asíncrono**: La investigación se ejecuta en segundo plano\n    - **Optimización paralela**: Múltiples sub-tareas ejecutadas simultáneamente  \n    - **Múltiples fuentes**: Integración con Tavily y Saptiva APIs\n    - **Alta disponibilidad**: Continúa funcionando aunque algunas APIs no estén disponibles",
        "operationId": "start_research_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Tarea de investigación aceptada y en ejecución",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                "example": {
                  "task_id": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "accepted",
                  "details": "Research task has been accepted and is running with optimized parallel processing."
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de solicitud inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/status": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Consultar estado de tarea",
        "description": "Obtiene el estado actual de una tarea de investigación específica",
        "operationId": "get_task_status_tasks__task_id__status_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de la tarea recuperado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                "example": {
                  "task_id": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "completed",
                  "details": "Research completed with 15 evidence sources"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{task_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Obtener reporte de investigación",
        "description": "Recupera el resultado completo de una tarea de investigación incluyendo el reporte y fuentes",
        "operationId": "get_report_reports__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte recuperado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Report"
                },
                "example": {
                  "status": "completed",
                  "report_md": "# Análisis del Mercado de IA en 2025\n\n## Resumen Ejecutivo\n...",
                  "sources_bib": "Generated from 15 evidence sources",
                  "metrics_json": "{\"mock_metric\": 1.0}"
                }
              }
            }
          },
          "404": {
            "description": "Reporte no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/traces/{task_id}": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "Obtener trazas de observabilidad",
        "description": "Recupera artefactos de trazabilidad y observabilidad para una tarea específica.\n    \n    Incluye:\n    - **Manifiesto**: Configuración y metadatos de la ejecución\n    - **Eventos**: Log de eventos en formato NDJSON para análisis\n    - **OpenTelemetry**: Trazas de rendimiento y monitoreo",
        "operationId": "get_traces_traces__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trazas recuperadas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Traces"
                },
                "example": {
                  "manifest_json": "{\"version\": \"0.7.0\", \"task_type\": \"research\", \"started_at\": \"2025-09-12T10:00:00Z\"}",
                  "events_ndjson": "{\"event\": \"task_started\", \"timestamp\": \"2025-09-12T10:00:00Z\"}\n{\"event\": \"plan_created\", \"subtasks\": 5}",
                  "otel_export_json": "{\"trace_id\": \"abc123\", \"spans\": [{\"name\": \"research_task\", \"duration_ms\": 1250}]}"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/deep-research": {
      "post": {
        "tags": [
          "deep-research"
        ],
        "summary": "Iniciar investigación profunda iterativa",
        "description": "Inicia una investigación profunda usando el patrón Together AI con refinamiento iterativo.\n    \n    ### 🔄 Características de Deep Research:\n    - **Refinamiento iterativo**: Múltiples ciclos de análisis y mejora\n    - **Score de completitud**: Evaluación automática de calidad\n    - **Gap analysis**: Identificación automática de brechas de información\n    - **Control de presupuesto**: Límites configurables de recursos\n    \n    ### 📊 Métricas de Calidad:\n    - Completion level (0.0 - 1.0)\n    - Research quality score \n    - Evidence count tracking\n    - Execution time monitoring",
        "operationId": "start_deep_research_deep_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeepResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Tarea de investigación profunda aceptada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                "example": {
                  "task_id": "deep-550e8400-e29b-41d4-a716-446655440000",
                  "status": "accepted",
                  "details": "Deep research task accepted with parallel processing. Configuration: 5 iterations, 0.85 min score."
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de configuración inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Consulta de investigación",
            "example": "Análisis del mercado de inteligencia artificial en 2025"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope",
            "description": "Alcance específico de la investigación",
            "example": "mercado_latinoamericano"
          },
          "budget": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget",
            "description": "Presupuesto máximo para fuentes de datos",
            "example": 50.0
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ResearchRequest",
        "description": "Solicitud para investigación estándar con procesamiento paralelo optimizado."
      },
...
    }
  },
...
}